Plot density between X et Y with rbokeh

This function take somes arguments from rbokeh and make plot.

plotXY(
  data,
  x,
  y,
  precision = 30,
  sizeOnCount = FALSE,
  outLine = TRUE,
  transform = NULL
)

Arguments

data
data.frame can be antaresData object

x
character, x variable

y
character, y variable

precision
numeric precision for plot

sizeOnCount
boolean, should addapt size of object based on count

outLine
boolean, add outline on your shape

transform
funciton, transform function apply on count (by cells), can be log

Examples

if (FALSE) {

setSimulationPath("myStudy")
myData <- readAntares()

plotXY(myData, "NODU", "LOAD", precision = 50,
         sizeOnCount = FALSE)

myData <- readAntares(areas = "all", links = "all")
myData <- mergeAllAntaresData(myData)
plotXY(myData, "OP. COST_max_b", "OP. COST_max_c", precision = 50,
         sizeOnCount = FALSE)

}