Get and set interactivity mode

setInteractivity globally sets the interactivity mode of plot functions. This is useful to avoid repeating interactive = FALSE or interactive = TRUE in each function. getInteractivity gets the interactivity mode.

setInteractivity(interactive = "auto")

getInteractivity()

Arguments

interactive
Should plot functions generate a UI that lets users to interactively modify input data and graphical parameters of a chart? It should be TRUE or FALSE. The default behavior is to set it to TRUE if the R session is interactive and to FALSE otherwise (for instance in Rmarkdown document).

Value

getInteractivity returns a boolean indicating the interactivity mode of plot functions. setInteractivity is only used for its side effects.